+++ TeamCode/src/main/java/org/betastudio/ftc/entry/InitializeRequested.java
@@ -1,4 +1,4 @@
-package org.betastudio.ftc.specification;
+package org.betastudio.ftc.entry;
 
 public interface InitializeRequested {
 	void init();
