+++ TeamCode/src/main/java/org/firstinspires/ftc/cores/structure/RotateOp.java
@@ -3,9 +3,9 @@ package org.firstinspires.ftc.cores.structure;
 import androidx.annotation.NonNull;
 
 import org.betastudio.ftc.action.Action;
-import org.betastudio.ftc.specification.HardwareController;
-import org.betastudio.ftc.specification.InitializeRequested;
-import org.betastudio.ftc.specification.TagOptionsRequired;
+import org.betastudio.ftc.entry.HardwareController;
+import org.betastudio.ftc.entry.InitializeRequested;
+import org.betastudio.ftc.entry.TagOptionsRequired;
 import org.firstinspires.ftc.teamcode.HardwareDatabase;
 import org.betastudio.ftc.util.Labeler;
 import org.firstinspires.ftc.teamcode.controllers.ServoCtrl;
