+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/controllers/ServoCtrl.java
@@ -5,7 +5,7 @@ import androidx.annotation.NonNull;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.betastudio.ftc.action.Action;
-import org.betastudio.ftc.specification.DashboardCallable;
+import org.betastudio.ftc.entry.DashboardCallable;
 import org.betastudio.ftc.ui.telemetry.TelemetryItem;
 import org.betastudio.ftc.util.message.TelemetryMsg;
